+++ tests/unittests/test_cli.py
@@ -65,7 +65,7 @@ class CLITest(unittest.TestCase):
         os.environ['COLUMNS'] = "120"  # Fake that the terminal is wide enough.
         self.assertRaises(SystemExit, next_action)
         self.assertEqual(call("""\
-usage: next-action [-h] [--version] [-c <config.cfg> | -C] [-f <todo.txt>] [-n <number> | -a] [-o] [-p [<priority>]]
+usage: next-action [-h] [--version] [-c [<config.cfg>]] [-f <todo.txt>] [-n <number> | -a] [-o] [-p [<priority>]]
 [<context|project> ...]
 
 Show the next action in your todo.txt. The next action is selected from the tasks in the todo.txt file based on task
@@ -76,9 +76,9 @@ optional arguments:
   -h, --help            show this help message and exit
   --version             show program's version number and exit
   --write-config-file   generate a sample configuration file and exit
-  -c <config.cfg>, --config-file <config.cfg>
-                        filename of configuration file to read (default: ~/.next-action.cfg)
-  -C, --no-config-file  don't read the configuration file
+  -c [<config.cfg>], --config-file [<config.cfg>]
+                        filename of configuration file to read (default: ~/.next-action.cfg); omit filename to not
+                        read any configuration file
   -f <todo.txt>, --file <todo.txt>
                         filename of todo.txt file to read; can be '-' to read from standard input; argument can be
                         repeated to read tasks from multiple todo.txt files (default: ~/todo.txt)